+++ packages/headless/src/features/standalone-search-box-set/standalone-search-box-set-actions.ts
@@ -43,6 +43,21 @@ export const registerStandaloneSearchBox = createAction(
     })
 );
 
+export interface ResetStandaloneSearchBoxActionCreatorPayload {
+  /**
+   * The standalone search box id.
+   */
+  id: string;
+}
+
+export const resetStandaloneSearchBox = createAction(
+  'standaloneSearchBox/reset',
+  (payload: ResetStandaloneSearchBoxActionCreatorPayload) =>
+    validatePayload(payload, {
+      id: requiredNonEmptyString,
+    })
+);
+
 export interface UpdateAnalyticsToSearchFromLinkActionCreatorPayload {
   /**
    * The standalone search box id.