+++ packages/headless/src/test/mock-insight-state.ts
@@ -1,6 +1,11 @@
 import {getCaseContextInitialState} from '../features/case-context/case-context-state';
 import {getConfigurationInitialState} from '../features/configuration/configuration-state';
+import {getFacetOptionsInitialState} from '../features/facet-options/facet-options-state';
+import {getCategoryFacetSetInitialState} from '../features/facets/category-facet-set/category-facet-set-state';
+import {getFacetSearchSetInitialState} from '../features/facets/facet-search-set/specific/specific-facet-search-set-state';
 import {getFacetSetInitialState} from '../features/facets/facet-set/facet-set-state';
+import {getNumericFacetSetInitialState} from '../features/facets/range-facets/numeric-facet-set/numeric-facet-set-state';
+import {getDateFacetSetInitialState} from '../features/facets/range-facets/date-facet-set/date-facet-set-state';
 import {getInsightConfigurationInitialState} from '../features/insight-configuration/insight-configuration-state';
 import {getInsightInterfaceInitialState} from '../features/insight-interface/insight-interface-state';
 import {getPaginationInitialState} from '../features/pagination/pagination-state';
@@ -27,8 +32,13 @@ export function buildMockInsightState(
     facetSet: getFacetSetInitialState(),
     pagination: getPaginationInitialState(),
     resultPreview: getResultPreviewInitialState(),
+    facetSearchSet: getFacetSearchSetInitialState(),
+    facetOptions: getFacetOptionsInitialState(),
     querySet: getQuerySetInitialState(),
     querySuggest: getQuerySuggestSetInitialState(),
+    numericFacetSet: getNumericFacetSetInitialState(),
+    dateFacetSet: getDateFacetSetInitialState(),
+    categoryFacetSet: getCategoryFacetSetInitialState(),
     ...config,
   };
 }