+++ packages/atomic/CHANGELOG.md
@@ -3,6 +3,77 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.82.3](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.82.2...@coveo/atomic@1.82.3) (2022-07-08)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+## [1.82.2](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.82.1...@coveo/atomic@1.82.2) (2022-07-08)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+## [1.82.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.82.0...@coveo/atomic@1.82.1) (2022-07-08)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+# [1.82.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.81.0...@coveo/atomic@1.82.0) (2022-07-07)
+
+
+### Features
+
+* **atomic,headless:** add search box to insights ([#2188](https://github.com/coveo/ui-kit/issues/2188)) ([d0465f6](https://github.com/coveo/ui-kit/commit/d0465f6a8fff4028f70485d4e0349840a34b6ed7))
+* **atomic:** hide arrow buttons ([#2175](https://github.com/coveo/ui-kit/issues/2175)) ([742e8dc](https://github.com/coveo/ui-kit/commit/742e8dcc1eceebffa590ca5ca30b1d1fe0f21ec4)), closes [#2179](https://github.com/coveo/ui-kit/issues/2179) [#2189](https://github.com/coveo/ui-kit/issues/2189)
+
+
+
+
+
+# [1.81.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.80.1...@coveo/atomic@1.81.0) (2022-07-07)
+
+
+### Bug Fixes
+
+* **atomic:** improve style of instant components in commerce example … ([#2193](https://github.com/coveo/ui-kit/issues/2193)) ([0f0945d](https://github.com/coveo/ui-kit/commit/0f0945d2c45300cb411b8a0d20c42729ae1046e2))
+
+
+### Features
+
+* npm workspace migration ([#2180](https://github.com/coveo/ui-kit/issues/2180)) ([f3b6997](https://github.com/coveo/ui-kit/commit/f3b69977675837dbed5ebe1eb3cb7fb2c98305ea))
+
+
+
+
+
+## [1.80.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.80.0...@coveo/atomic@1.80.1) (2022-07-06)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+# [1.80.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.79.1...@coveo/atomic@1.80.0) (2022-07-06)
+
+
+### Features
+
+* **headless:** Add analytics actions to insight-search ([#2147](https://github.com/coveo/ui-kit/issues/2147)) ([e5c72ad](https://github.com/coveo/ui-kit/commit/e5c72ad5c2073dd7ebeb3afc087246735c99a4e2))
+
+
+
+
+
 ## [1.79.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.79.0...@coveo/atomic@1.79.1) (2022-07-05)